+++ CHANGELOG.md
@@ -4,7 +4,9 @@
 
 ### Fixed
 
+- #115 - Add default `status` and `sequence` values for iCal notifications missing these fields
 - #124 - Handle encoded non-ASCII characters in email subjects.
+- #126 - Ignore a class of non-maintenance-notification emails from Telia.
 
 ## v2.0.7 - 2021-12-01